[AIFlow] Move logging config to settings.py
