+++ lib/notification_service/tests/test_server_config.py
@@ -18,7 +18,7 @@
 import os
 import unittest
 
-from notification_service.server_config import NotificationServerConfig, get_configuration, get_configuration_file_path
+from notification_service.server_config import NotificationServerConfig
 
 
 class TestNotificationServerConfig(unittest.TestCase):
@@ -31,35 +31,6 @@ class TestNotificationServerConfig(unittest.TestCase):
         self.assertEqual('sqlite:///ns.db', ns_config.db_uri)
         self.assertFalse(ns_config.enable_ha)
         self.assertEqual(10000, ns_config.ha_ttl_ms)
-
-    def test_get_configuration(self):
-        os.environ['NOTIFICATION_HOME'] = os.path.dirname(__file__)
-        try:
-            ns_config = get_configuration()
-            self.assertEqual(50052, ns_config.port)
-            self.assertEqual('127.0.0.1:50052', ns_config.advertised_uri)
-            self.assertEqual('sqlite:///ns.db', ns_config.db_uri)
-            self.assertFalse(ns_config.enable_ha)
-            self.assertEqual(10000, ns_config.ha_ttl_ms)
-        finally:
-            del os.environ['NOTIFICATION_HOME']
-
-    def test_get_configuration_path(self):
-        prev_home = os.environ['HOME']
-        try:
-            os.environ['HOME'] = "/non_exist_dir"
-            with self.assertRaises(FileNotFoundError):
-                get_configuration_file_path()
-            os.environ['NOTIFICATION_HOME'] = "/non_exist_dir"
-            with self.assertRaises(FileNotFoundError):
-                get_configuration_file_path()
-
-            current_dir = os.path.dirname(__file__)
-            os.environ['NOTIFICATION_HOME'] = current_dir
-            self.assertEqual(os.path.join(current_dir, "notification_server.yaml"), get_configuration_file_path())
-        finally:
-            os.environ['HOME'] = prev_home
-            del os.environ['NOTIFICATION_HOME']
 
 
 if __name__ == '__main__':
